pyproject.toml
[tool.poetry]
name = "clipper"
version = "0.1.0"
description = ""
authors = ["Frank Zhang <fzhang0723@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
flask = "^3.1.0"
flask-cors = "^5.0.0"
pydantic = {extras = ["email"], version = "^2.10.2"}
pip = "^24.3.1"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
rq = "^2.0.0"
yt-dlp = "^2024.11.18"
weaviate-client = "^4.9.6"
torch = "^2.5.1"
transformers = "^4.47.0"
opencv-python = "^4.10.0.84"
pillow = "^10.0.0"
imageio-ffmpeg = "^0.5.1"
numpy = "^2.1.3"
imageio = "^2.36.1"
proglog = "^0.1.10"
moviepy = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
